+++ packages/f-header/CHANGELOG.md
@@ -3,6 +3,19 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v2.0.0-beta-8
+------------------------------
+*September 20, 2019*
+
+### Changed
+ - Props in `Navigation.vue`
+
+### Added
+ - New logic to return URL for login and logout
+ - Added new computed function to get `userInfo`
+ - Axios for GET request
+
+
 v2.0.0-beta-7
 ------------------------------
 *September 18, 2019*